another commit

+++ src/main/java/com/example/springdocker/model/MyMathCalc.java
@@ -15,7 +15,7 @@ public class MyMathCalc {
         return a * b;
     }
 
-    public double divide(int a, int b){
+    public int divide(int a, int b){
         return a / b;
     }
 }
